deploy/deploy_restAPI.py
from flask import Flask, request, jsonify
import joblib

app = Flask(__name__)
model = joblib.load('recommender_model.pkl')

@app.route('/predict', methods=['POST'])
def predict():
    customer_id = request.json['customer_id']
    # Add logic to generate recommendations
    recommendations = generate_recommendations(customer_id)
    return jsonify(recommendations)

def generate_recommendations(customer_id):
    # Implement recommendation logic
    return []

if __name__ == '__main__':
    app.run(debug=True)
