pyproject.toml
[tool.poetry]
name = "recommendation-system"
version = "0.1.0"
description = "A recommendation system to suggest products and quantities for businesses using MLflow and Surprise."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.0"
numpy = "^1.21.0"
scikit-learn = "^1.5.0"
scikit-surprise = "^1.1.4"
mlflow = "^2.1.0"
flask = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.0"
black = "^21.5b0"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
